# api/openapi.yaml
openapi: 3.0.3
info:
  title: General Scheduler API
  version: 1.0.0
servers:
  - url: https://api.brand.com
paths:
  /v1/dates:
    get:
      summary: List future dates with availability
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date
        - in: query
          name: days
          schema:
            type: integer
            default: 30
            maximum: 90
        - in: query
          name: location_id
          schema:
            type: string
        - in: query
          name: person_id
          schema:
            type: string
        - in: query
          name: timezone
          schema:
            type: string
      responses:
        '200':
          description: OK
  /v1/slots:
    get:
      summary: Get exposed slots (2–5 random) for a date
      parameters:
        - in: query
          name: date
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: location_id
          required: true
          schema:
            type: string
        - in: query
          name: person_id
          schema:
            type: string
        - in: query
          name: timezone
          schema:
            type: string
      responses:
        '200':
          description: OK
  /v1/bookings:
    post:
      summary: Place a hold and create booking
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - slot_id
                - customer
              properties:
                slot_id:
                  type: string
                customer:
                  type: object
                  required:
                    - name
                    - phone
                  properties:
                    name:
                      type: string
                    phone:
                      type: string
                    email:
                      type: string
                notes:
                  type: string
                consent:
                  type: object
                source:
                  type: string
      responses:
        '201':
          description: Created
        '409':
          description: Conflict
  /v1/bookings/{id}/confirm:
    post:
      summary: Confirm a held booking
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /admin/v1/availabilities:
    post:
      summary: Create or update availability rule
      responses:
        '201':
          description: Created
  /admin/v1/slots/generate:
    post:
      summary: Generate slot instances from rules
      responses:
        '200':
          description: OK
  /admin/v1/bookings:
    get:
      summary: List bookings
      responses:
        '200':
          description: OK
  /admin/v1/bookings/{id}:
    patch:
      summary: Cancel or reschedule a booking
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK